Correct return type

+++ include/ForgeScanUtils/timing_utils.h
@@ -25,7 +25,7 @@ public:
         m_bRunning = false;
     }
     
-    double elapsedMilliseconds()
+    long long elapsedMilliseconds()
     {
         std::chrono::time_point<std::chrono::system_clock> endTime;
